Fixed missing function reference
