+++ src/ReplaysLolRedditBot/Bot.py
@@ -1,12 +1,11 @@
 import json
 from time import sleep
-import os
 import logging
 from bs4 import BeautifulSoup
 from dotenv import load_dotenv
 from Errors import SubmissionExistsException
 from Helpers import load_config_from_local_directory, get_praw_client_from_config, get_matches_from_link, \
-    is_submission_id_present_in_list_of_dictionaries, configure_logger
+    is_submission_id_present_in_list_of_dictionaries, configure_logger, get_submission_file_path, initialise_submissions
 
 
 def main():
@@ -18,7 +17,7 @@ def main():
     configure_logger()
     reddit = get_praw_client_from_config(configuration)
     submission_file_path = get_submission_file_path()
-    initialise_submissions(submission_file_path)
+    scraped_submissions = initialise_submissions(submission_file_path)
     while True:
         try:
             for submission in reddit.subreddit(target_subreddit).new(limit=submission_limit):
@@ -47,12 +46,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
-
-def initialise_submissions(submission_file_path):
-    if os.path.exists(submission_file_path):
-        with open(submission_file_path, "r") as jsonfile:
-            scraped_submissions = json.load(jsonfile)
-    else:
-        scraped_submissions = []
-    return scraped_submissions
